code clean-ups (#101)

+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/parts/GoldBar.java
@@ -22,7 +22,7 @@ public class GoldBar extends Animation {
     private static final Random RANDOM = new SecureRandom();
 
     private final double goldDiffToSpawnCoin;
-    private final int millisecondsForGoldCount;
+    private final long millisecondsForGoldCount;
 
     private Double lastGold;
     private LocalTime lastGoldCheck;
